// packages/ui/src/components/picker/picker.types.ts
export interface PositionGroup {
  readonly category: string;
  readonly positions: readonly string[];
}